Reverting the Revert: 7a029eff9561da31191723530c5854851a534b39

+++ sugarcrm/ModuleInstall/ModuleInstaller.php
@@ -215,10 +215,13 @@ class ModuleInstaller{
 				$rac = new RepairAndClear();
 				$rac->repairAndClearAll($selectedActions, $installed_modules,true, false);
 				$this->rebuild_relationships();
-				$this->log('<br><b>' . translate('LBL_MI_COMPLETE') . '</b>');
-
 				UpdateSystemTabs('Add',$tab_modules);
-
+				
+				//clear the unified_search_module.php file 
+	            require_once('modules/Home/UnifiedSearchAdvanced.php');
+	            UnifiedSearchAdvanced::unlinkUnifiedSearchModulesFile(); 
+	            				
+				$this->log('<br><b>' . translate('LBL_MI_COMPLETE') . '</b>');
 		}else{
 			die("No \$installdefs Defined In $this->base_dir/manifest.php");
 		}
@@ -328,7 +331,9 @@ class ModuleInstaller{
 				//if it's not a sugar file then we remove it otherwise we can't restor it
 				if(!$this->ms->sugarFileExists($to)){
 					$GLOBALS['log']->debug('ModuleInstaller[uninstall_new_file] deleting file ' . $to);
-					unlink($to);
+					if(file_exists($to)) {
+					    unlink($to);
+					}
 				}else{
 					$GLOBALS['log']->fatal('ModuleInstaller[uninstall_new_file] Could not remove file ' . $to . ' as no backup file was found to restore to');
 				}
@@ -1252,6 +1257,10 @@ class ModuleInstaller{
 
 				UpdateSystemTabs('Restore',$installed_modules);
 
+	            //clear the unified_search_module.php file 
+	            require_once('modules/Home/UnifiedSearchAdvanced.php');
+	            UnifiedSearchAdvanced::unlinkUnifiedSearchModulesFile();     				
+				
 				$this->log('<br><b>' . translate('LBL_MI_COMPLETE') . '</b>');
 				if(!$this->silent){
 					update_progress_bar('install', $total_steps, $total_steps);
@@ -1280,6 +1289,14 @@ class ModuleInstaller{
 			$this->merge_files('Ext/Vardefs/', 'vardefs.ext.php');
 			sugar_cache_reset();
 	}
+    //BEGIN SUGARCRM flav=pro ONLY
+    function rebuild_dependencies(){
+            $this->log(translate('LBL_MI_REBUILDING') . " Dependencies...");
+			$this->merge_files('Ext/Dependencies/', 'deps.ext.php');
+			sugar_cache_reset();
+	}
+    //END SUGARCRM flav=pro ONLY
+
 	function rebuild_layoutdefs(){
             $this->log(translate('LBL_MI_REBUILDING') . " Layoutdefs...");
 			$this->merge_files('Ext/Layoutdefs/', 'layoutdefs.ext.php');
@@ -1293,12 +1310,12 @@ class ModuleInstaller{
 
 	function rebuild_dashletcontainers(){
             $this->log(translate('LBL_MI_REBUILDING') . " DC Actions...");
-			$this->merge_files('Ext/DashletContainer/Containers', 'dcactions.ext.php');
+			$this->merge_files('Ext/DashletContainer/Containers/', 'dcactions.ext.php');
 	}
 
 	function rebuild_modules(){
             $this->log(translate('LBL_MI_REBUILDING') . " Modules...");
-			$this->merge_files('Ext/Include', 'modules.ext.php', '', true);
+			$this->merge_files('Ext/Include/', 'modules.ext.php', '', true);
 	}
 
 	function rebuild_administration(){
@@ -1353,6 +1370,9 @@ class ModuleInstaller{
 
 		$this->rebuild_languages($sugar_config['languages']);
 		$this->rebuild_vardefs();
+        //BEGIN SUGARCRM flav=pro ONLY
+        $this->rebuild_dependencies();
+        //END SUGARCRM flav=pro ONLY
 		$this->rebuild_layoutdefs();
 		$this->rebuild_menus();
 		$this->rebuild_dashletcontainers();
@@ -1391,18 +1411,14 @@ class ModuleInstaller{
 						     if (substr($entry, 0, 9) == '_override') {
 						    	$override[] = $entry;
 						    } else {
-							    $fp = sugar_fopen($module_install . '/' . $entry, 'r');
-							    $file = fread($fp , filesize($module_install . '/' . $entry));
+							    $file = file_get_contents($module_install . '/' . $entry);
 							    $GLOBALS['log']->debug(get_class($this)."->merge_files(): found {$module_install}{$entry}") ;
-							    fclose($fp);
 							    $extension .= "\n". str_replace(array('<?php', '?>', '<?PHP', '<?'), array('','', '' ,'') , $file);
 						    }
 						}
 					}
 					foreach ($override as $entry) {
-						$fp = sugar_fopen($module_install . '/' . $entry, 'r');
-                        $file = fread($fp , filesize($module_install . '/' . $entry));
-                        fclose($fp);
+                        $file = file_get_contents($module_install . '/' . $entry);
                         $extension .= "\n". str_replace(array('<?php', '?>', '<?PHP', '<?'), array('','', '' ,'') , $file);
 					}
 				}
@@ -1437,9 +1453,7 @@ class ModuleInstaller{
 								if((empty($filter) || substr_count($entry, $filter) > 0) && is_file($module_install.'/'.$entry)
 								  && $entry != '.' && $entry != '..' && strtolower(substr($entry, -4)) == ".php")
 								{
-									$fp = sugar_fopen($module_install . '/' . $entry, 'r');
-									$file = fread($fp , filesize($module_install . '/' . $entry));
-									fclose($fp);
+									$file = file_get_contents($module_install . '/' . $entry);
 									$extension .= "\n". str_replace(array('<?php', '?>', '<?PHP', '<?'), array('','', '' ,'') , $file);
 								}
 						}
@@ -1620,6 +1634,7 @@ private function dir_get_files($path, $base_path){
 private function dir_file_count($path){
 	//if its a file then it has at least 1 file in the directory
 	if(is_file($path)) return 1;
+	if(!is_dir($path)) return 0;
 	$d = dir($path);
 	$count = 0;
 	while ($e = $d->read()){
@@ -1988,7 +2003,7 @@ private function dir_file_count($path){
                         $path ='custom/Extension/' . $relationship['module']. '/Ext/Layoutdefs';
                     }
 				}
-                
+
 				if(!empty($relationship['module_layoutdefs']) && file_exists($path . '/'. $this->id_name . '.php')){
 					mkdir_recursive($path . '/'.DISABLED_PATH, true);
 					rename( $path . '/'. $this->id_name . '.php', $path . '/'.DISABLED_PATH.'/'. $this->id_name . '.php');
