+++ sugarcrm/tests/modules/Administration/Bug36978Test.php
@@ -6,13 +6,10 @@ var $rel_guid;
 var $has_custom_table_dictionary;	
 var $moduleList;
 
-function setUp() {
-	
-    if(true) {
-       $this->markTestSkipped("Skipping unless otherwise specified");
-       return;
-    }	
-	
+function setUp() 
+{
+    $this->markTestSkipped("Skipping unless otherwise specified");
+    
     $admin = new User();
     $GLOBALS['current_user'] = $admin->retrieve('1');	
 	
@@ -234,17 +231,33 @@ function tearDown() {
        unlink('modules/abc_Test/metadata/studio.php'); 
     }    
     
-	rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Vardefs');
-	rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Layoutdefs');
-	rmdir_recursive('custom/Extension/modules/abc_Test/Ext');
-	rmdir_recursive('custom/Extension/modules/abc_Test');
-	rmdir_recursive('modules/abc_Test/metadata');
-	rmdir_recursive('modules/abc_Test');
+    if(is_dir('custom/Extension/modules/abc_Test/Ext/Vardefs')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Vardefs');
+    }
+    if(is_dir('custom/Extension/modules/abc_Test/Ext/Layoutdefs')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Layoutdefs');
+    }
+    if(is_dir('custom/Extension/modules/abc_Test/Ext')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test/Ext');
+    }
+    if(is_dir('custom/Extension/modules/abc_Test')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test');
+    }
+    if(is_dir('modules/abc_Test/metadata')) {
+        rmdir_recursive('modules/abc_Test/metadata');
+    }
+    if(is_dir('modules/abc_Test')) {
+        rmdir_recursive('modules/abc_Test');
+    }
 	
-	$sql = "DELETE FROM relationships WHERE id = '{$this->rel_guid}'";
-	$GLOBALS['db']->query($sql);
+    if ( !empty($this->rel_guid) ) {
+        $sql = "DELETE FROM relationships WHERE id = '{$this->rel_guid}'";
+        $GLOBALS['db']->query($sql);
+	}
 	
-	$GLOBALS['moduleList'] = $this->moduleList;
+	if ( !empty($this->moduleList) ) {
+        $GLOBALS['moduleList'] = $this->moduleList;
+    }
 }
